core: fixup iterator type to resolve nitty GCC warning

+++ src/solver/core.cc
@@ -49,7 +49,7 @@ solver::node* solver::solve(const std::vector<size_t>& solution_indexes, int dep
     // lower average (i.e. narrow down the solutions more) as being more
     // favourable.
     const auto best_guesses = get_best_unique_match_guesses(solution_indexes);
-    const int num_guesses = best_guesses.size();
+    const size_t num_guesses = best_guesses.size();
 
     // Allocate new nodes. It is more efficient to dynamically allocate these
     // contiguously.
@@ -58,7 +58,7 @@ solver::node* solver::solve(const std::vector<size_t>& solution_indexes, int dep
     // Traverse the decision tree further for the top candidates, spreading the
     // work over threads if appropriate.
     std::vector<std::thread> threads;
-    for (int i = 0; i < num_guesses; i++) {
+    for (size_t i = 0; i < num_guesses; i++) {
         const size_t guess_index = best_guesses[i].second;
         node* subtree = &subtrees[i];
         subtree->guess_index = guess_index;
